requirements.txt
streamlit
mysql-connector-python
pymongo
google-generativeai
python-dotenv
pandas